requirements.txt

Django==4.2.11
djangorestframework==3.14.0
django-timezone-field==6.1.0
django-cors-headers
requests==2.32.3
psycopg2
django-extensions==3.2.3

pytest
pytest-django
pytest-cov
psycopg2-binary



